requirements.txt
astroid==4.0.1
dill==0.4.0
guizero==1.6.0
isort==7.0.0
mccabe==0.7.0
pillow==12.0.0
platformdirs==4.5.0
pylint==4.0.2
tomlkit==0.13.3
